fix: remove flags field
Helps https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/29896

The flags field is set as an unsigned integer, which is not compatible
with InfluxDB 1.x. Luckily, the field is also not useful, as it stores
a value chosen from an enum containing no useful values.

In this change, the flags field is removed from conversion. If it needs
to be added in the future, we can expect that enough time will have
passed that retention policies will have removed the prior `flags` field.