fix(influx2otel): only map to histogram when all fields match
When AddPoint is called with the unknown metric type, it tries to
infer the type from the given field names. Right now it infers histogram
when any of the fields matches the histogram keys, but that leads to a
lot of errors when your data isn't actually a histogram. Instead, only
infer a histogram if all fields match the expected histogram keys.